requirements.txt
torch
torchvision
hydra-core
lightning
einops
torch-fidelity
webdataset
gradio
transformers
scikit-learn
wandb
timm
diffusers